# Fixed AVC444v1 luma UV averaging