The old implementation of AVC444v1 did not use averaging for luma UV values.
This pull fixes that.

performance results for `1920x1080:`
* old (without averaging)
```
PROFILER NAME            |    COUNT |       TOTAL |  AVG      |    IPS
RGBToYUV444-ssse3        |      500 |     0.8615s | 0.001723s |    580
RGBToYUV444-BGRX         |      500 |     4.3929s | 0.008786s |    114
RGBToYUV444-generic      |      500 |     8.8536s | 0.017707s |     56
```

* new (with averaging)
```
PROFILER NAME            |    COUNT |       TOTAL |  AVG      |    IPS
RGBToYUV444-ssse3-avg    |      500 |     0.8341s | 0.001668s |    599
RGBToYUV444-BGRX-avg     |      500 |     5.0998s | 0.010200s |     98
RGBToYUV444-generic-avg  |      500 |     9.4429s | 0.018886s |     53
```