@nfedera 2018-01-31
Please post performance values of both, the SSSE3 and the generic function form before and the code in this pr. Thanks.