@nfedera 2018-01-31
Data looks good now, nice job